do correction to movie collection assignment part 2
do correction to movie collection assignment part 2 by adding
authentication functionality and checking permissions.
Start final example project for book_review.

+++ movie_collection/movies/views.py
@@ -1,5 +1,7 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.urls import reverse
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
 
 from .models import Movie
 from .forms import MovieForm
@@ -9,6 +11,7 @@ def movie_list(request):
     movies = Movie.objects.all()
     return render(request, "movies/movie-list.html", {"movies": movies})
 
+@login_required
 def create_movie(request):
     form = MovieForm()
     if request.method == "POST":
@@ -16,6 +19,8 @@ def create_movie(request):
         if form.is_valid():
             form.save()
             return redirect("movies:movie_list")
+        else:
+            print(form.errors)
 
     context = {"form": form}
     return render(request, "movies/create-movie.html", context)
@@ -24,10 +29,16 @@ def movie_detail(request, movie_pk):
     movie = get_object_or_404(Movie, pk=movie_pk)
     return render(request, "movies/movie-detail.html", {"movie": movie})
 
+@login_required
 def edit_movie(request, movie_pk):
     movie = get_object_or_404(Movie, pk=movie_pk)
     form = MovieForm(instance=movie)
     if request.method == "POST":
+        if request.user != movie.added_by:
+            messages.error(request, "You do not have permission to edit that movie")
+            return redirect("movies:movie_list")
+        
+
         form = MovieForm(request.POST, request.FILES, instance=movie)
         if form.is_valid():
             form.save()
@@ -35,12 +46,21 @@ def edit_movie(request, movie_pk):
     context = {"form": form, "movie": movie}
     return render(request, "movies/edit-movie.html", context)
 
+@login_required
 def confirm_delete(request, movie_pk):
     movie = get_object_or_404(Movie, pk=movie_pk)
+    if request.user != movie.added_by:
+        messages.error(request, "You do not have permission to delete that movie")
+        return redirect("movies:movie_list")
     return render(request, "movies/confirm-delete.html", {"movie": movie})
 
+@login_required
 def delete_movie(request, movie_pk):
     movie = get_object_or_404(Movie, pk=movie_pk)
+    if request.user != movie.added_by:
+        messages.error(request, "You do not have permission to delete that movie")
+        return redirect("movies:movie_list")
+    
     if request.method == "POST":
         movie.delete()
         return redirect("movies:movie_list")
